deployments: use "--tail 10" to follow logs
docker-compose uses "--tail all" by default, which prints the complete
logs before starting to follow new lines. This becomes a problem when
the service has been running for a lot of time and there are thousand of
lines.

Use "--tail 10" to mimic the behavior of the "tail" program, which shows
the latest 10 lines by default.